+++ src-tauri/src/lib.rs
@@ -13,6 +13,7 @@ use core::state::AppState;
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
     tauri::Builder::default()
+        .plugin(core::log::plugin())
         .plugin(tauri_plugin_opener::init())
         .plugin(tauri_plugin_clipboard_x::init())
         .plugin(tauri_plugin_store::Builder::new().build())